chore: merge trivial test files into neighboring test modules
- tests/speech/test_system_integration.py merged into tests/speech/test_config.py
- tests/telemetry/test_telemetry_record.py merged into tests/telemetry/test_store.py
