+++ tests/speech/test_config.py
@@ -17,3 +17,10 @@ def test_jarvis_config_has_speech():
     assert hasattr(cfg, "speech")
     assert isinstance(cfg.speech, SpeechConfig)
     assert cfg.speech.backend == "auto"
+
+
+def test_jarvis_system_has_speech_backend():
+    """JarvisSystem has a speech_backend attribute."""
+    from openjarvis.system import JarvisSystem
+
+    assert "speech_backend" in JarvisSystem.__dataclass_fields__
